+++ src/printer/json_printer.hpp
@@ -28,18 +28,18 @@ namespace mtconnect::printer {
     JsonPrinter(uint32_t jsonVersion, const std::string version = "", bool pretty = false);
     ~JsonPrinter() override = default;
 
-    std::string printErrors(const unsigned int instanceId, const unsigned int bufferSize,
+    std::string printErrors(const uint64_t instanceId, const unsigned int bufferSize,
                             const uint64_t nextSeq, const ProtoErrorList &list) const override;
 
-    std::string printProbe(const unsigned int instanceId, const unsigned int bufferSize,
+    std::string printProbe(const uint64_t instanceId, const unsigned int bufferSize,
                            const uint64_t nextSeq, const unsigned int assetBufferSize,
                            const unsigned int assetCount, const std::list<DevicePtr> &devices,
                            const std::map<std::string, size_t> *count = nullptr) const override;
 
-    std::string printSample(const unsigned int instanceId, const unsigned int bufferSize,
+    std::string printSample(const uint64_t instanceId, const unsigned int bufferSize,
                             const uint64_t nextSeq, const uint64_t firstSeq, const uint64_t lastSeq,
                             observation::ObservationList &results) const override;
-    std::string printAssets(const unsigned int anInstanceId, const unsigned int bufferSize,
+    std::string printAssets(const uint64_t anInstanceId, const unsigned int bufferSize,
                             const unsigned int assetCount,
                             const asset::AssetList &asset) const override;
     std::string mimeType() const override { return "application/mtconnect+json"; }
